feat: 선착순 이벤트 플래그 초기화 로직 추가
